+++ docs/CHANGELOG.md
@@ -3,6 +3,8 @@
 # Next
 
 * Package `types/tensors/numpy` with methods to read and write tensors from/to `.npy` and `.npz` files.
+* Package `simplego`:
+  * Fixed bug introduced in parallelize version of Erf(x).
 
 # v0.19.5: 2024/05/30 SimpleGo (go) backend optimizations