+++ docs/CHANGELOG.md
@@ -2,6 +2,11 @@
 
 # Next
 
+* Package `backends`:
+  * Method `New()` and `NewWithConfig()` will be changed to return an error (as opposed to panic).
+    Temporarily the methods `MustNew()` (which panics on errors, like today) and `NewOrErr` (which returns
+    an error) were created to have a clear API, and `New()` was marked as deprecated. At the next version
+    `New()` will change the API.
 * Package `types/tensors/numpy` with methods to read and write tensors from/to `.npy` and `.npz` files.
 * Package `simplego`:
   * Fixed bug introduced in parallelize version of Erf(x).